Update lib structure
