+++ src/lib/T3000/Hvac/Basic/Basic.Line.ts
@@ -1,5 +1,5 @@
 
-import Path from './Path';
+import Path from './Basic.Path';
 
 class Line extends Path {
