+++ src/lib/T3000/Hvac/Basic/Basic.ts
@@ -1,0 +1,8 @@
+
+class Basic {
+  constructor() {
+    console.log('Basic');
+  }
+}
+
+export default Basic;
