--- src/lib/T3000/Hvac/Basic/Graphics.ts
@@ -1,8 +1,0 @@
-
-class Graphics {
-  constructor() {
-    console.log('Graphics');
-  }
-}
-
-export default Graphics;
