+++ src/lib/T3000/Hvac/Data/Data.Constant.ts
