--- src/lib/T3000/Hvac/Data/Point.ts
@@ -1,13 +1,0 @@
-
-class Point {
-
-  public x: number;
-  public y: number;
-
-  constructor(x: number, y: number) {
-    this.x = x || 0;
-    this.y = y || 0;
-  }
-}
-
-export default Point;
