+++ src/lib/T3000/Hvac/Doc/ContentHeader.ts
@@ -1,0 +1,18 @@
+
+import { PageRecord, Point } from '../Model/Model';
+
+class ContentHeader {
+  public Page: PageRecord;
+  public MaxWorkDim: Point;
+
+  constructor() {
+    this.Initialize();
+  }
+
+  Initialize = () => {
+    this.Page = new PageRecord();
+    this.MaxWorkDim = new Point(32e4, 32e4);
+  }
+}
+
+export default ContentHeader;
