+++ src/lib/T3000/Hvac/Doc/Line.ts
@@ -1,6 +1,6 @@
 
-import Models from '../Model/Hvac.Models';
-import * as Utils from '../Helper/Hvac.Utils';
+import Models from '../Data/Data.Constant';
+import * as Utils from '../Helper/Helper.Utils';
 import Point from '../Data/Point';
 
 class Line {
