+++ src/lib/T3000/Hvac/Helper/Helper.SVG1.ts
